requirements_v3.txt
requests
aiogram>=3.0.0
python-dotenv